+++ CHANGES.md
@@ -1,2 +1,3 @@
 - Add option to not auto enable experimentation mode when opening tracked recipe
 - Ignore pre dragonflight crafting tools
+- Fix optimization target menu title